Temporary fix for bug 47954 - keep section link in edit summary
Since talk page edits don't necessary belong to a seciton (or may
involve several sections), we'll need to be more clever in the
DiscussionParser to extract it for the long term solution. This
temporary solution will expose any section link in the edit summary
payload so that people can jump straight to the section if they
want to (same as existing behavior in the archive version).

Bug: 47954
Change-Id: I2745ba194ba1f9b5b7c446588da6586e87d35b31