ext.echo.badge: Remove defunct CSS rule
The space after colon on this line is actually a non-breaking space
(&nbsp;), which is not allowed in this place according to CSS syntax.
Therefore all browsers ignored this rule (logging an error like
"Invalid CSS property value") and the shadow was never shown.

We could fix the rule instead, but I'm pretty sure users are used to
this by now, and in my opinion it looks better without shadow.

Bug: 53490
Change-Id: I1fd508d2059bec5cd79a6dcdce8dc9be6e6d4229